Rename createLocation to enhanceLocation
